Split up rust bridge
