+++ src/rust/Cargo.toml
@@ -77,18 +77,39 @@ tracy-client = { version = "=0.17.0", features = [
 # the build and about 1.5MB of object code to the resulting binary. So we want
 # to keep the number of hosts down when and if it's easy to, but having the
 # number grow gradually is also not the end of the world.
+#
+
+# Note: due to some subtle mis-features in Cargo's unification of transitive
+# dependency versions, we do _not_ refer to all the separate soroban crate
+# versions directly here. See $(top_srcdir)/src/Makefile.am for the logic that
+# builds each crate separately and then provides them to the stellar-core crate
+# here.
+#
+# However this fact makes it difficult to use an IDE to work on this crate since
+# the IDE will not be able to resolve the hosts to anything at all. So for the sake
+# of making IDE-edits, we keep some commented-out copies of the host dependencies
+# here. These are not used by the build system, but if you uncomment them while
+# working they will allow your IDE to at least find "nearly correct" versions of
+# the host dependencies. Make sure they are commented back out before committing
+# (and be careful to reset any changes the IDE makes to Cargo.lock)
+
+# [dependencies.soroban-env-host-p23]
+# version = "=23.0.0"
+# git = "https://github.com/stellar/rs-soroban-env"
+# package = "soroban-env-host"
+# rev = "31cb455b87a25a1c049360b5422c411deae63ba2"
 
 # [dependencies.soroban-env-host-p22]
 # version = "=22.0.0"
 # git = "https://github.com/stellar/rs-soroban-env"
 # package = "soroban-env-host"
-# rev = "0497816694bef2b103494c8c61b7c8a06a72c7d3"
+# rev = "1cd8b8dca9aeeca9ce45b129cd923992b32dc258"
 
 # [dependencies.soroban-env-host-p21]
-# version = "=21.2.0"
+# version = "=21.2.2"
 # git = "https://github.com/stellar/rs-soroban-env"
 # package = "soroban-env-host"
-# rev = "8809852dcf8489f99407a5ceac12625ee3d14693"
+# rev = "7eeddd897cfb0f700f938b0c8d6f0541150d1fcb"
 
 # The test wasms and synth-wasm crate should usually be taken from the highest
 # supported host, since test material usually just grows over time.
